[Modify] 카테고리 컴포넌트 데이터 통신 완료
