+++ src/pages/Product/Product.js
@@ -12,6 +12,12 @@ export class Product extends Component {
     super();
     this.state = {
       productCard: [],
+      filterState: {
+        bone: false,
+        hair: false,
+        growth: false,
+        skin: false,
+      },
     };
   }
 
@@ -25,6 +31,54 @@ export class Product extends Component {
         });
       });
   }
+
+  makeCondition = () => {
+    const filterMatch = {
+      bone: 1,
+      hair: 2,
+      growth: 3,
+      skin: 4,
+    };
+
+    const filtered = Object.entries(this.state.filterState).reduce(
+      (acc, [key, value]) => {
+        if (!acc && value) {
+          return acc + `efficacy=${filterMatch[key]}`;
+        }
+
+        if (value) {
+          return acc + `&efficacy=${filterMatch[key]}`;
+        }
+        return acc;
+      },
+      ''
+    );
+
+    fetch(`${GET_PRODUCTS_API}?${filtered}`)
+      .then(res => res.json())
+      .then(data => {
+        this.setState({
+          productCard: data.message,
+        });
+      });
+    console.log(`filtered>>>>>`, filtered);
+  };
+
+  handleCheckBox = event => {
+    const checkBoxName = event.target.name;
+    const checkBoxNameState = !this.state.filterState[checkBoxName];
+    this.setState(
+      {
+        filterState: {
+          ...this.state.filterState,
+          [checkBoxName]: checkBoxNameState,
+        },
+      },
+      () => {
+        this.makeCondition();
+      }
+    );
+  };
 
   render() {
     const { productCard } = this.state;
@@ -42,7 +96,11 @@ export class Product extends Component {
           <h2 className="sr-only">Product Body</h2>
 
           <form className="productCategory">
-            <ProductCategory key={productCard.id} />
+            <ProductCategory
+              key={productCard.id}
+              makeCondition={this.makeCondition}
+              handleCheckBox={this.handleCheckBox}
+            />
           </form>
 
           <ul className="productList">
