+++ src/pages/Product/ProductCategory/ProductCategory.js
@@ -1,81 +1,8 @@
 import React, { Component } from 'react';
-import { GET_PRODUCTS_API } from '../../../config';
 
 import './ProductCategory.scss';
 
 export class ProductCategory extends Component {
-  constructor() {
-    super();
-    this.state = {
-      filterState: {
-        bone: false,
-        hair: false,
-        growth: false,
-        skin: false,
-      },
-    };
-  }
-
-  componentDidMount() {
-    // fetch(`${GET_PRODUCTS_API}`)
-    fetch(`${GET_PRODUCTS_API}`)
-      .then(res => res.json())
-      .then(data => {
-        this.setState({
-          productCard: data.message,
-        });
-      });
-  }
-
-  makeCondition = () => {
-    const filterMatch = {
-      bone: 1,
-      hair: 2,
-      growth: 3,
-      skin: 4,
-    };
-
-    const filtered = Object.entries(this.state.filterState).reduce(
-      (acc, [key, value]) => {
-        if (!acc && value) {
-          return acc + `efficacy=${filterMatch[key]}`;
-        }
-
-        if (value) {
-          return acc + `&efficacy=${filterMatch[key]}`;
-        }
-        return acc;
-      },
-      ''
-    );
-
-    // fetch(`${GET_PRODUCTS_API}?${filtered}`);
-    fetch(`${GET_PRODUCTS_API}?efficacy=1`)
-      .then(res => res.json())
-      .then(data => {
-        this.setState({
-          productCard: data.message,
-        });
-      });
-    console.log(filtered);
-  };
-
-  handleCheckBox = event => {
-    const checkBoxName = event.target.name;
-    const checkBoxNameState = !this.state.filterState[checkBoxName];
-    this.setState(
-      {
-        filterState: {
-          ...this.state.filterState,
-          [checkBoxName]: checkBoxNameState,
-        },
-      },
-      () => {
-        this.makeCondition();
-      }
-    );
-  };
-
   render() {
     return PRODUCT_CATEGORY.map(category => {
       return (
@@ -83,7 +10,7 @@ export class ProductCategory extends Component {
           <input
             type="checkbox"
             name={category.name}
-            onChange={this.handleCheckBox}
+            onChange={this.props.handleCheckBox}
           />
           <label>
             <img className="iconImage" alt={category.alt} src={category.src} />
